Added some comments/documentation to most files. And refractored some naming conventions. Also moved UserParser from use cases pacakge to API package
