+++ src/main/java/Controllers/RunCommand.java
@@ -19,7 +19,7 @@ public class RunCommand {
      */
     public RunCommand(int command) {
         UserAnalyzerFactory factory = new UserAnalyzerFactory();
-        this.userAnalyzer = factory.create_analyzer(command);
+        this.userAnalyzer = factory.createAnalyzer(command);
     }
 
     /**
