+++ src/main/java/Entities/IExercise.java
@@ -2,6 +2,9 @@ package Entities;
 
 import java.util.ArrayList;
 
+/**
+ *  This is interface is an abstraction of Exercise Entity
+ */
 public interface IExercise {
 
     String getName();
